% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germlineStrelka2Manta.R
\name{germlineStrelka2Manta}
\alias{germlineStrelka2Manta}
\title{germlineStrelka2Manta}
\usage{
germlineStrelka2Manta(
  bam,
  ref = "/gpfs42/projects/lab_lpasquali/shared_data/marc/ref/hg38/hg38.fa",
  out_dir,
  cores,
  conf_manta = "configManta.py",
  conf_strelka2 = "configureStrelkaGermlineWorkflow.py",
 
    call_regions = "/gpfs42/projects/lab_lpasquali/shared_data/marc/ref/hg38/GRCh38_exclude_decoys_small_contigs.bed"
)
}
\arguments{
\item{bam}{BAM file to use in the analysis}

\item{ref}{Reference fasta genome. By default '/gpfs42/projects/lab_lpasquali/shared_data/marc/ref/hg38/hg38.fa'}

\item{out_dir}{Path where the output of the analysis will be saved.}

\item{cores}{Number of cores to use.}

\item{conf_manta}{Manta 'configManta.py' path. By default 'configManta.py'}

\item{conf_strelka2}{Strelka2 'configureStrelkaGermlineWorkflow.py' path. By default 'configureStrelkaGermlineWorkflow.py'.}

\item{call_regions}{Bed file of all the chromosomes that should be included in the analysis.
For instance, the following bed file could be provided to exclude all decoys and small contigs.
By default '/gpfs42/projects/lab_lpasquali/shared_data/marc/ref/hg38/GRCh38_exclude_decoys_small_contigs.bed'}
}
\description{
Germline SNVs and SVs calling using Strelka2 and Manta
}
\examples{
\dontrun{}
}
