% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pealignBwa.R
\name{pealignBwa}
\alias{pealignBwa}
\title{pealignBwa}
\usage{
pealignBwa(
  fq1,
  fq2,
 
    ref = "/gpfs42/projects/lab_lpasquali/shared_data/ref/GCA_000001405.15_GRCh38_no_alt_analysis_set.fa",
  out_dir,
  threads = 1,
  temp = file.path(out_dir, "temp"),
  platform = "ILLUMINA",
  bwa = "bwa",
  samtools = "samtools",
  samblaster = "samblaster"
)
}
\arguments{
\item{fq1}{Fastq R1 file of paired-end data.}

\item{fq2}{Fastq R2 file of paired-end data.}

\item{ref}{Reference genome (ref.fa) to use for the alignment.}

\item{threads}{Number of threads to use in the analysis.}

\item{platform}{Platforme used in the sequecing process. By default 'ILLUMINA'.}

\item{bwa}{Path to the executable bwa. By default 'bwa'.}

\item{samtools}{Path to the executable bwa. By default 'samtools'.}

\item{samblaster}{Path to the executable bwa. By default 'samblaster'.}
}
\description{
Align paired-end fq files using bwa mem
}
\examples{
\dontrun{
fq1 <- '/media/msubirana/insulinomas/epimutations/processed/hg37/bam/longranges/wgs/sub_10759_R1.fa'
fq2 <- '/media/msubirana/insulinomas/epimutations/processed/hg37/bam/longranges/wgs/sub_10759_R2.fa'
ref <- '/home/msubirana/Documents/phd/refgen/hg38/hg38.fa'
out_dir <- '/media/msubirana/insulinomas/epimutations/processed/hg37/bam/longranges/wgs/subset_proves'
threads <- 6
pealignBwa(fq1=fq1,
           fq2=fq2,
           ref=ref,
           out_dir=out_dir)
}
}
