% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realignBwa.R
\name{realignBwa}
\alias{realignBwa}
\title{realignBwa}
\usage{
realignBwa(
  bam,
  ref,
  out_dir,
  threads = 1,
  temp = file.path(out_dir, "temp"),
  platform = "ILLUMINA",
  bwa = "bwa",
  samtools = "samtools",
  samblaster = "samblaster"
)
}
\arguments{
\item{bam}{Paired-end BAM file to realign}

\item{ref}{Reference genome (ref.fa) to use for the alignment.}

\item{threads}{Number of threads to use in the analysis.}

\item{platform}{Platforme used in the sequecing process. By default 'ILLUMINA'.}

\item{bwa}{Path to the executable bwa. By default 'bwa'.}

\item{samtools}{Path to the executable bwa. By default 'samtools'.}

\item{samblaster}{Path to the executable bwa. By default 'samblaster'.}
}
\description{
Convert paired-end bam files to fastqs and then realign with bwa mem
}
\examples{
\dontrun{
bam <- '/media/msubirana/insulinomas/epimutations/processed/hg37/bam/longranges/wgs/sub_10759.bam'
ref <- '/home/msubirana/Documents/phd/refgen/hg38/hg38.fa'
out_dir <- '/media/msubirana/insulinomas/epimutations/processed/hg37/bam/longranges/wgs/subset_proves'
threads <- 6
realignBwa(bam=bam,
           ref=ref,
           out_dir=out_dir)
}
}
