% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germlineStrelka2Manta.R
\name{strelka2}
\alias{strelka2}
\title{strelka2}
\usage{
strelka2(
  bam,
  ref,
  out_dir,
  cores,
  conf_strelka2 = "configManta.py",
 
    call_regions = "/gpfs42/projects/lab_lpasquali/shared_data/marc/ref/hg38/GRCh38_exclude_decoys_small_contigs.bed",
  indel_candidates
)
}
\arguments{
\item{bam}{BAM file to use in the analysis}

\item{ref}{Reference fasta genome. By default '/gpfs42/projects/lab_lpasquali/shared_data/marc/ref/hg38/hg38.fa'}

\item{out_dir}{Path where the output of the analysis will be saved.}

\item{cores}{Number of cores to use.}

\item{conf_strelka2}{Strelka2 'configureStrelkaGermlineWorkflow.py' path. By default 'configureStrelkaGermlineWorkflow.py'.}

\item{call_regions}{Bed file of all the chromosomes that should be included in the analysis.
For instance, the following bed file could be provided to exclude all decoys and small contigs.
By default '/gpfs42/projects/lab_lpasquali/shared_data/marc/ref/hg38/GRCh38_exclude_decoys_small_contigs.bed'}

\item{indel_candidates}{'candidateSmallIndels.vcf.gz' file from Manta calling. It is a recommended best practice to provide indel candidates from the Manta SV and indel caller.}
}
\description{
SVs calling of paired tumor-normal wgs using strelka2
}
\examples{
\dontrun{}
}
